rag_system/requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
openai==1.98.0
tiktoken==0.9.0
numpy>=1.26.2
python-magic-bin==0.4.14
tqdm==4.67.1
pathspec==0.12.1